summary of changes from 0.9 to 0.10
The classes `info._InfoJsonDecoder` and `registers._RelationalInfoJsonDecoder`, introduced in version 0.6 and now removed, had a very strange bug that caused LMDB to throw `lmdb.ReadersFullError` if `registers.Register.apos` was called too many times. I do not understand why this bug occurred, but I was able to isolate and remove it.

Essentially, if the `object_hook` field of `json.JSONDecoder` pointed to either a method of the `json.JSONDecoder` object OR to a lambda function, the error would be thrown after sufficiently many `registers.Register.apos` calls. Therefore, it was difficult to pass an `lmdb.Transaction` to the `object_hook` function, which was necessary for `registers._InfoJsonDecoder`.

My solution is to, essentially, manually reimplement the bugged portion of `json.JSONDecoder.decode`. The relevant changes are the following:

- info._Info._default_str_hook
- info._Info.from_json
- registers.Register._RelationalApriInfoStrHook

DATABASE CHANGES:
- slightly abbreviated apos values and apri values

OTHER CHANGES
- removed `info._Info.set_encoder` and `info._Info.clean_encoder`.
- changed `registers.Register._check_blk_open_raise` to instance-method